refactor(*): 테스트 케이스 클래스 별 파일 분리
